src/requirements.txt
prompt_toolkit
requests
python-dotenv
Pygments
